fix(runtime): keep gated_attention sigmoid gate in float32
The ROCm reference gated_attention lane reassigned the sigmoid gate via
`1.0 / (1.0 + np.exp(-g))`, which numpy's >=2.2 type stubs model as
float64 even though it stays float32 at runtime — a local mypy false
positive (CI pins numpy<2.2). Make the float32 explicit; value-identical,
and consistent with the surrounding `.astype(np.float32)` idiom.

+++ python/tessera/runtime.py
@@ -11403,7 +11403,7 @@ def _execute_rocm_compiled_exotic_attention(artifact: RuntimeArtifact,
         g = np.asarray(gate).astype(np.float32)
         act = str(kwargs.get("gate_activation", "sigmoid"))
         if act == "sigmoid":
-            g = 1.0 / (1.0 + np.exp(-g))
+            g = (1.0 / (1.0 + np.exp(-g))).astype(np.float32)
         elif act not in ("identity", "none"):
             raise ValueError(
                 "gate_activation must be 'sigmoid', 'identity', or 'none'")
